feat: add production capacity query by available stock

+++ src/main/java/com/stockcontrol/RawMaterial/services/ProductService.java
@@ -5,7 +5,9 @@ import com.stockcontrol.RawMaterial.exceptions.BarcodeExistException;
 import com.stockcontrol.RawMaterial.exceptions.CodeExistException;
 import com.stockcontrol.RawMaterial.models.ProductModel;
 import com.stockcontrol.RawMaterial.models.StockModel;
+import com.stockcontrol.RawMaterial.repositories.ProductCrudRepository;
 import com.stockcontrol.RawMaterial.repositories.ProductRepository;
+import com.stockcontrol.RawMaterial.view.ProductCapacityView;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,6 +20,7 @@ import java.util.Optional;
 public class ProductService {
 
     final ProductRepository productRepository;
+    final ProductCrudRepository productCrudRepository;
     final StockService stockService;
 
     @Transactional
@@ -47,11 +50,20 @@ public class ProductService {
 
     @Transactional
     public Optional<ProductModel> delete(Optional<ProductModel> product0){
-        productRepository.delete(product0.get());
+
+        if (product0.isPresent()){
+            productRepository.delete(product0.get());
+            stockService.delete(product0.get().getCode());
+        }
         return product0;
     }
 
     public List<ProductModel> allProducts(){
         return productRepository.findAll();
     }
+
+    public List<ProductCapacityView> productionvailable(){
+        return productCrudRepository.findProductionCapacity();
+    }
+
 }
